+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.9.5...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.10.0...HEAD) - TBD
+
+## [4.10.0](https://github.com/schemathesis/schemathesis/compare/v4.9.5...v4.10.0) - 2026-02-07
 
 ### :rocket: Added